# node/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documenting API BattleTime with Swagger.",
        "version": "0.2.0",
        "title": "API Documented"
    },
    "host": "localhost:8000",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/users/image_identification": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Change front and bakc identification image url",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [{
                    "name": "front_image",
                    "in": "formData",
                    "description": "Image front identification",
                    "required": true,
                    "type": "file"
                    },
                    {
                        "name": "back_image",
                        "in": "formData",
                        "description": "Image back identification",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [{
                    "auth_token": [
                        "write:user",
                        "read:user"
                    ]
                }]
            }
        },
...
    },
    "securityDefinitions": {
        "auth_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}